organising files, adding more arg checks and tests, list of all models
